Edit pt channel, server, category + fix delete message on delete server

+++ BizChat/Controllers/ServersController.cs
@@ -3,6 +3,7 @@ using BizChat.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using Newtonsoft.Json.Serialization;
 
 namespace BizChat.Controllers
 {
@@ -74,9 +75,55 @@ namespace BizChat.Controllers
 				}
 				// cam asa ar trebui sa facem verificarile
 				ViewBag.IsModerator = db.ServerUsers.Where(su => su.ServerId == serverId && su.UserId == _userManager.GetUserId(User)).First().IsModerator;
+				ViewBag.IsOwner = db.ServerUsers.Where(su => su.ServerId == serverId && su.UserId == _userManager.GetUserId(User)).First().IsOwner;
+				ViewBag.UserRole = db.Roles.Find(db.UserRoles.Where(u => u.UserId == _userManager.GetUserId(User)).First().RoleId).Name;
 				// in view verificam ViewBag.IsModerator == true
 			}
 			return View();
+		}
+
+		[HttpPost]
+		[Authorize(Roles = "RegisteredUser, AppModerator, AppAdmin")]
+		public IActionResult Index([FromForm] Category category, int? serverId, int? channelId, int? e_categoryId)
+		{
+			string UserId = _userManager.GetUserId(User);
+			var servers = db.Servers.Where(s => s.Users.Where(su => su.UserId == UserId).Count() > 0);
+			ViewBag.Servers = servers;
+			if (serverId != null && _signInManager.IsSignedIn(User))
+			{
+				ViewBag.ServerId = serverId;
+
+				var channels = db.Channels.Where(channel => channel.ServerId == serverId);
+				ViewBag.Channels = channels;
+
+				var servermembers = db.ServerUsers.Where(su => su.ServerId == serverId);
+				ViewBag.ServerMembers = db.ApplicationUsers.
+										Where(user => servermembers.Where(sm => sm.UserId == user.Id).First() != null);
+
+				IQueryable<Category>? categories = db.Categories.Where(c => c.ServerId == serverId);
+				ViewBag.Categories = categories;
+
+				if (channelId != null)
+				{
+					Channel? selectedChannel = db.Channels.Find(channelId);
+					ViewBag.selectedChannel = selectedChannel;
+				}
+				var is_user_mod = db.ServerUsers.Where(su => su.ServerId == serverId && su.UserId == _userManager.GetUserId(User)).First().IsModerator;
+				ViewBag.IsModerator = is_user_mod;
+				ViewBag.IsOwner = db.ServerUsers.Where(su => su.ServerId == serverId && su.UserId == _userManager.GetUserId(User)).First().IsOwner;
+				ViewBag.UserRole = db.Roles.Find(db.UserRoles.Where(u => u.UserId == _userManager.GetUserId(User)).First().RoleId).Name;
+
+				if (is_user_mod == true)
+				{
+					if (e_categoryId != null && ModelState.IsValid)
+					{
+						Category category_to_edit = db.Categories.Find(e_categoryId);
+						category_to_edit.CategoryName = category.CategoryName;
+						db.SaveChanges();
+					}
+				}
+			}
+			return RedirectToAction(actionName: "Index", controllerName: "Servers", routeValues: new { serverId, channelId });
 		}
 
 		[Authorize(Roles = "RegisteredUser, AppModerator, AppAdmin")]
@@ -110,7 +157,42 @@ namespace BizChat.Controllers
 			}
 		}
 
-		[Authorize(Roles ="RegisteredUser, AppModerator, AppAdmin")]
+		[Authorize(Roles = "RegisteredUser, AppModerator, AppAdmin")]
+		public IActionResult EditServer(int serverId)
+		{
+			if (db.ServerUsers.Where(su => su.UserId == _userManager.GetUserId(User) && su.ServerId == serverId).First().IsOwner == true)
+			{
+				Server server = db.Servers.Find(serverId);
+				return View(server);
+			}
+			else
+			{
+				return RedirectToAction("Index");
+			}
+		}
+
+		[HttpPost]
+		[Authorize(Roles = "RegisteredUser, AppModerator, AppAdmin")]
+		public IActionResult EditServer(Server newserver)
+		{
+			Server server = db.Servers.Find(newserver.Id);
+			if (ModelState.IsValid)
+			{
+				if (db.ServerUsers.Where(su => su.UserId == _userManager.GetUserId(User) && su.ServerId == server.Id).First().IsOwner == true)
+				{
+					server.Name = newserver.Name;
+					server.Description = newserver.Description;
+					db.SaveChanges();
+				}
+				return RedirectToAction(actionName: "Index", controllerName: "Servers", routeValues: new { serverId = server.Id });
+			}
+			else
+			{
+				return View(server);
+			}
+		}
+
+		[Authorize(Roles = "RegisteredUser, AppModerator, AppAdmin")]
 		public IActionResult NewCategory(int serverId)
 		{
 			Console.WriteLine(serverId);
@@ -134,7 +216,7 @@ namespace BizChat.Controllers
 			{
 				db.Categories.Add(category);
 				db.SaveChanges();
-				return RedirectToAction(actionName: "Index", routeValues: new { serverId = category.ServerId});
+				return RedirectToAction(actionName: "Index", routeValues: new { serverId = category.ServerId });
 			}
 			else
 			{
@@ -155,7 +237,7 @@ namespace BizChat.Controllers
 		[Authorize(Roles = "RegisteredUser, AppModerator, AppAdmin")]
 		public IActionResult NewChannel(Channel channel)
 		{
-			if(ModelState.IsValid)
+			if (ModelState.IsValid)
 			{
 				db.Add(channel);
 				db.SaveChanges();
@@ -168,12 +250,42 @@ namespace BizChat.Controllers
 		}
 
 		[HttpPost]
-		[Authorize(Roles = "AppModerator, AppAdmin")]
+		[Authorize(Roles = "RegisteredUser, AppModerator, AppAdmin")]
+		public IActionResult EditChannel([FromForm]Channel channel)
+		{
+			if(ModelState.IsValid)
+			{
+				db.Channels.Find(channel.Id).Name = channel.Name;
+				db.Channels.Find(channel.Id).Description= channel.Description;
+				db.SaveChanges();
+			}
+			return RedirectToAction(actionName: "Index", routeValues: new { serverId = channel.ServerId, channelId = channel.Id });
+		}
+
+		[NonAction]
+		public void DeleteChannelMessages(int channelId)
+		{
+			db.Messages.RemoveRange(db.Messages.Where(m => m.ChannelId == channelId));
+		}
+
+		[HttpPost]
+		[Authorize(Roles = "RegisteredUser, AppModerator, AppAdmin")]
 		public IActionResult DeleteServer(int serverId)
 		{
+			List<string> high_role_ids = (from r in db.Roles where r.Name != "RegisteredUser" select r.Id).ToList();
 			// Delete Server with the serverId
-			db.Servers.Remove(db.Servers.Find(serverId)!);
-			db.SaveChanges();
+			var user_id = _userManager.GetUserId(User);
+			var user_role_in_server = db.ServerUsers.Where(su => su.UserId == user_id && su.ServerId == serverId).First();
+			if (user_role_in_server.IsOwner == true || db.UserRoles.Where(ur => ur.UserId == user_id && high_role_ids.Contains<string>(ur.RoleId)).First() != null)
+			{
+				db.Servers.Remove(db.Servers.Find(serverId)!);
+				var s_channels = db.Channels.Where(c => c.ServerId == serverId);
+				foreach (var channel in s_channels)
+				{
+					DeleteChannelMessages(channel.Id);
+				}
+				db.SaveChanges();
+			}
 			return RedirectToAction(controllerName: "Servers", actionName: "Index");
 		}
 
@@ -188,7 +300,7 @@ namespace BizChat.Controllers
 			message.UserId = _userManager.GetUserId(User);
 			db.Messages.Add(message);
 			db.SaveChanges();
-			return Json(new { Message = "Success"});
+			return Json(new { Message = "Success" });
 		}
 	}
 }
